Switch license to AGPLv3+
The license was initially MPLv2 as I expected the project to be a fairly
simple tiny web app, but it has already grown into multiple files. Now
that I'm going to add a server component, I'm switching the license for
future versions to AGPL like all my other non-library projects.